refactor(botonic-react): rename file to open-persistence-menu and leave only one component inside
