+++ packages/botonic-react/src/webchat/components/opened-persistent-menu.jsx
@@ -5,9 +5,9 @@ import LogoMenu from '../../assets/menuButton.svg'
 import { Button } from '../../components/button'
 import { ROLES, WEBCHAT } from '../../constants'
 import { WebchatContext } from '../../contexts'
-import { ConditionalAnimation } from '../components/conditional-animation'
 import { useComponentVisible } from '../hooks'
 import { Icon } from './common'
+import { ConditionalAnimation } from './conditional-animation'
 
 const ButtonsContainer = styled.div`
   position: absolute;
@@ -57,32 +57,6 @@ export const OpenedPersistentMenu = ({ onClick, options, borderRadius }) => {
         </ButtonsContainer>
       )}
     </div>
-  )
-}
-
-export const PersistentMenu = ({ onClick, persistentMenu }) => {
-  const { getThemeProperty } = useContext(WebchatContext)
-
-  const persistentMenuOptions = getThemeProperty(
-    WEBCHAT.CUSTOM_PROPERTIES.persistentMenu,
-    persistentMenu
-  )
-
-  const CustomMenuButton = getThemeProperty(
-    WEBCHAT.CUSTOM_PROPERTIES.customMenuButton,
-    undefined
-  )
-
-  return (
-    <>
-      {persistentMenuOptions ? (
-        <ConditionalAnimation>
-          <div role={ROLES.PERSISTENT_MENU_ICON} onClick={onClick}>
-            {CustomMenuButton ? <CustomMenuButton /> : <Icon src={LogoMenu} />}
-          </div>
-        </ConditionalAnimation>
-      ) : null}
-    </>
   )
 }
